+++ Intern/rayx-core/src/Debug/Instrumentor.h
@@ -28,7 +28,7 @@ namespace RAYX {
 struct RAYX_API ProfileResult {
     std::string Name;
     long long Start, End;
-    uint32_t ThreadID;
+    size_t ThreadID;
 };
 
 struct RAYX_API InstrumentationSession {
@@ -110,7 +110,7 @@ class RAYX_API InstrumentationTimer {
         long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
         long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-        uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+        size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
         Instrumentor::Get().WriteProfile({m_Name, start, end, threadID});
 
         if (m_canPrint && BENCH_FLAG) {
